{{cookiecutter.repo_name}}/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
--find-links https://prototyp3-dev.github.io/pip-wheels-riscv/wheels/

certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
cytoolz==0.12.2
    # via eth-utils
eth-abi==4.2.1
    # via python-cartesi
eth-hash==0.5.2
    # via eth-utils
eth-typing==3.5.2
    # via
    #   eth-abi
    #   eth-utils
eth-utils==2.3.1
    # via eth-abi
exceptiongroup==1.2.0
    # via pytest
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
packaging==23.2
    # via pytest
parsimonious==0.9.0
    # via eth-abi
pluggy==1.3.0
    # via pytest
pycryptodome==3.19.0
    # via
    #   -r requirements.in
    #   python-cartesi
pydantic==1.10.13
    # via python-cartesi
pytest==7.4.4
    # via
    #   -r requirements.in
    #   python-cartesi
python-cartesi==0.1.0
    # via -r requirements.in
regex==2023.12.25
    # via parsimonious
requests==2.31.0
    # via
    #   -r requirements.in
    #   python-cartesi
tomli==2.0.1
    # via pytest
toolz==0.12.0
    # via cytoolz
typing-extensions==4.9.0
    # via
    #   eth-typing
    #   pydantic
urllib3==2.1.0
    # via requests
